refactor: rename file TextArea.tsx to Pane.tsx
