+++ src/components/Pane.tsx
